KNN implementation for ARduino Mega2560
This is a simple implementation of KNN algorithm for Arduino. In
knnarduino3,ino you will find a program that allows you classify data in
the format "{x,x,x,x}" where x are numbers in double format. In
knnarduino4.ino you can use "prueba.h" this is a database for get the
best k for classification.